assignment5/requirements.txt
h5py
scikit-learn
Pillow
scikit-image
numpy
